Q9 - I changed the totalPrice in Transaction.java to double and realized I can simplify calculateTotalPrice. I also added discount and vat at the receipt so I can crosscheck easily

+++ src/main/java/airlinereservationsystem/Transaction.java
@@ -6,9 +6,9 @@ public class Transaction {
     String destination;
     int ticketQuantity;
     String flightClass;
-    int totalPayment;
+    double totalPayment;
 
-    public Transaction(String name, String airline, String destination, int ticketQuantity, String flightClass, int totalPayment) {
+    public Transaction(String name, String airline, String destination, int ticketQuantity, String flightClass, double totalPayment) {
         this.name = name;
         this.airline = airline;
         this.destination = destination;
